Move SlotEventsChannel into ethereum namespace.
