+++ ethereum/events/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+  implementation project(':infrastructure:events')
+  testFixturesImplementation project(':infrastructure:unsigned')
+}
