+++ ethereum/events/src/main/java/tech/pegasys/teku/ethereum/events/SlotEventsChannel.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.util.time.channels;
+package tech.pegasys.teku.ethereum.events;
 
 import tech.pegasys.teku.infrastructure.events.VoidReturningChannelInterface;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
