+++ ethereum/statetransition/build.gradle
@@ -2,6 +2,7 @@ dependencies {
   implementation project(':bls')
   implementation project(':data')
   implementation project(':ethereum:core')
+  implementation project(':ethereum:events')
   implementation project(':ethereum:pow:api')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:weaksubjectivity')
