+++ services/beaconchain/build.gradle
@@ -4,6 +4,7 @@ dependencies {
   implementation project(':infrastructure:metrics')
   implementation project(':data:provider')
   implementation project(':ethereum:core')
+  implementation project(':ethereum:events')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:statetransition')
