+++ settings.gradle
@@ -4,6 +4,7 @@ include 'bls'
 include 'ethereum:core'
 include 'ethereum:spec'
 include 'ethereum:dataproviders'
+include 'ethereum:events'
 include 'ethereum:networks'
 include 'ethereum:pow:api'
 include 'ethereum:statetransition'
