+++ validator/coordinator/src/main/java/tech/pegasys/teku/validator/coordinator/performance/PerformanceTracker.java
@@ -14,11 +14,11 @@
 package tech.pegasys.teku.validator.coordinator.performance;
 
 import java.util.Set;
+import tech.pegasys.teku.ethereum.events.SlotEventsChannel;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.versions.altair.SyncCommitteeSignature;
-import tech.pegasys.teku.util.time.channels.SlotEventsChannel;
 
 public interface PerformanceTracker extends SlotEventsChannel {
